Add seven figures temporarily
An important figure's filter was fixed, also, it was
confirmed with the professor that the warning that
the shannon index figure does not need to be
addressed. Additionally figures representing
taxonomy by sample were added with their respective
taxonomy tables. Not all of these will be kept. Also
there are two figures in progress.