Add firgures and corrections to knit
The focus was to realign the purpose and to form a
story while trying to pass lintr and be able to
knit. All was completed. Remarks and changes by
Dr. Zimmerman were made and then modified. Figures
liike the shannon index and tables were modified to
better show the data.